requirements.txt
africastalking==1.1.9
asgiref==3.2.10
boto==2.49.0
boto3==1.9.130
botocore==1.12.217
bump2version==0.5.10
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
contextlib2==0.5.5
coverage==5.3
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.42.0
django-autocomplete-light==3.8.1
django-cors-headers==3.0.2
django-grappelli==2.13.1
django-modeltranslation==0.13.2
django-nested-admin==3.2.4
django-otp==1.0.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-smart-selects==1.5.4
django-smart-selects-711==1.0.0
django-storages==1.7.2
django-templated-mail==1.1.1
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
djoser==2.0.5
docutils==0.14
drf-writable-nested==0.6.2
geographiclib==1.49
geopy==1.20.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-logging==1.14.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
jmespath==0.9.4
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
Pillow==6.2.1
protobuf==3.10.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
schema==0.7.1
signals==0.0.2
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
utils==1.0.1
